+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 1.0.4
 
-* Support for flutter  3.13.5 and added a loading indicator
+* Support for flutter  3.13.6 and added transaction verification